feat : order 서비스 배치 프로세스 설정[날마다 정산으로 바꿔서 보내기]

+++ order/src/main/java/com/dev_high/order/presentation/dto/OrderRegisterRequest.java
@@ -7,7 +7,7 @@ public record OrderRegisterRequest(
     String sellerId,
     String buyerId,
     String auctionId,
-    Integer confirmAmount,
+    Long confirmAmount,
     LocalDateTime confirmDate,
     OrderStatus status
 ) {
